ref: Tests have been devided in client functions. Clean up fixtures after tests finish execution. Now is not required to build image every time that tests are run
