+++ tests/client/test_http_client.py
@@ -3,148 +3,7 @@ import requests
 
 from schema_registry.client import SchemaRegistryClient, load
 
-from tests.client import data_gen
-
-
-def assertLatest(self, meta_tuple, sid, schema, version):
-    self.assertNotEqual(sid, -1)
-    self.assertNotEqual(version, -1)
-    self.assertEqual(meta_tuple[0], sid)
-    self.assertEqual(meta_tuple[1], schema)
-    self.assertEqual(meta_tuple[2], version)
-
-
-def test_register(client):
-    parsed = load.loads(data_gen.BASIC_SCHEMA)
-    schema_id = client.register("test-basic-schema", parsed)
-
-    assert schema_id > 0
-    assert len(client.id_to_schema) == 1
-
-
-def test_register_json_data(client, deployment_schema):
-    schema_id = client.register("test-deployment", deployment_schema)
-    assert schema_id > 0
-
-
-def test_register_with_custom_headers(client, country_schema):
-    headers = {"custom-serialization": "application/x-avro-json"}
-    schema_id = client.register("test-country", country_schema, headers=headers)
-    assert schema_id > 0
-
-
-def test_multi_subject_register(client):
-    parsed = load.loads(data_gen.BASIC_SCHEMA)
-    schema_id = client.register("test-basic-schema", parsed)
-    assert schema_id > 0
-
-    # register again under different subject
-    dupe_id = client.register("test-basic-schema-backup", parsed)
-    assert schema_id == dupe_id
-    assert len(client.id_to_schema) == 1
-
-
-def test_dupe_register(client):
-    parsed = load.loads(data_gen.BASIC_SCHEMA)
-    subject = "test-basic-schema"
-    schema_id = client.register(subject, parsed)
-
-    assert schema_id > 0
-    latest = client.get_schema(subject)
-
-    # register again under same subject
-    dupe_id = client.register(subject, parsed)
-    assert schema_id == dupe_id
-
-    dupe_latest = client.get_schema(subject)
-    assert latest == dupe_latest
-
-
-def test_getters(client):
-    parsed = load.loads(data_gen.BASIC_SCHEMA)
-    subject = "subject-does-not-exist"
-    version = client.check_version(subject, parsed)
-    assert version is None
-
-    # There is already a registered schema with id 1
-    schema = client.get_by_id(1)
-    assert schema is not None
-
-    latest = client.get_schema(subject)
-    assert latest == (None, None, None, None)
-
-    # register
-    schema_id = client.register(subject, parsed)
-    schema = client.get_schema(subject, schema_id)
-    assert schema is not None
-
-    latest = client.get_schema(subject)
-    version = client.check_version(subject, parsed)
-    fetched = client.get_by_id(schema_id)
-
-    assert fetched == parsed
-
-
-def test_multi_register(client):
-    """
-    Register two different schemas under the same subject
-    with backwards compatibility
-    """
-    version_1 = load.loads(data_gen.USER_V1)
-    version_2 = load.loads(data_gen.USER_V2)
-    subject = "test-user-schema"
-
-    id1 = client.register(subject, version_1)
-    latest_schema_1 = client.get_schema(subject)
-    client.check_version(subject, version_1)
-
-    id2 = client.register(subject, version_2)
-    latest_schema_2 = client.get_schema(subject)
-    client.check_version(subject, version_2)
-
-    assert id1 != id2
-    assert latest_schema_1 != latest_schema_2
-    # ensure version is higher
-    assert latest_schema_1.version < latest_schema_2.version
-
-    client.register(subject, version_1)
-    latest_schema_3 = client.get_schema(subject)
-    # latest should not change with a re-reg
-    assert latest_schema_2 == latest_schema_3
-
-
-def test_compatibility(client, user_schema_v3):
-    """
-    Test the compatibility of a new User Schema against the latest one.
-    The last user schema is the V2.
-    """
-    compatibility = client.test_compatibility("test-user-schema", user_schema_v3)
-    assert compatibility
-
-
-def test_update_compatibility(client):
-    """
-    The latest User V2 schema is  BACKWARD and FORWARDFULL compatibility (FULL).
-    So, we can ipdate compatibility level for the specified subject.
-    """
-    assert client.update_compatibility("FULL", "test-user-schema") == "FULL"
-
-
-def test_get_compatibility(client):
-    """
-    Test latest compatibility for test-user-schema subject
-    """
-    assert client.get_compatibility("test-user-schema") == "FULL"
-
-
-def test_delete_subject(client, user_schema_v3):
-    subject = "subject-to-delete"
-    versions = [load.loads(data_gen.USER_V1), load.loads(data_gen.USER_V2)]
-
-    for version in versions:
-        client.register(subject, version)
-
-    assert client.delete_subject(subject) == [1, 2]
+from tests import data_gen
 
 
 def test_context(client):
